+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TopLevelBuild.java
@@ -789,7 +789,7 @@ public class TopLevelBuild extends BaseBuild {
 			Dom4JUtil.getNewElement(
 				"p", null, "Start Time: ",
 				JenkinsResultsParserUtil.toDateString(
-					new Date(getStartTimestamp()))),
+					new Date(getStartTime()), getJenkinsReportTimeZoneName())),
 			Dom4JUtil.getNewElement(
 				"p", null, "Build Time: ",
 				JenkinsResultsParserUtil.toDurationString(getDuration())),